[IMP] mail:tests: unskip 'chat window should remain folded when new message is received'
